+++ packages/terra-data-grid/tests/wdio/flowsheet-data-grid-spec.js
@@ -53,7 +53,7 @@ Terra.describeViewports('FlowsheetDataGrid', ['medium', 'large'], () => {
 
       it('renders a default flowsheet data grid', () => {
         browser.keys(['Tab']); // Cell 0,0 gets focus
-        expect(browser.$('tr.column-header-row th:nth-child(1)').isFocused()).toBe(true);
+        expect(browser.$('[class*="column-header-row"] th:nth-child(1)').isFocused()).toBe(true);
         Terra.validates.element('default-flowsheet-data-grid', { selector: '#default-terra-flowsheet-data-grid-table' });
       });
     });
@@ -79,6 +79,34 @@ Terra.describeViewports('FlowsheetDataGrid', ['medium', 'large'], () => {
 
         browser.keys(['ArrowUp', 'ArrowUp']);
         expect(browser.$('//*[@id="terra-flowsheet-data-grid-no-column-headers-table-rowheader-1"]').isFocused()).toBe(true);
+      });
+    });
+
+    describe('flowsheet data grid with column highlight', () => {
+      it('renders a flowsheet with default column highlight', () => {
+        browser.url('/raw/tests/cerner-terra-framework-docs/data-grid/flowsheet-data-grid/column-highlight-orange');
+        browser.keys(['Tab']); // Cell 0,0 gets focus
+        expect(browser.$('[class*="column-header-row"] th:nth-child(1)').isFocused()).toBe(true);
+        Terra.validates.element('flowsheet-data-grid-column-highlight', { selector: '#terra-flowsheet-data-grid-column-highlight-orange' });
+      });
+
+      it('renders a flowsheet data grid with green column highlight', () => {
+        browser.url('/raw/tests/cerner-terra-framework-docs/data-grid/flowsheet-data-grid/column-highlight-green');
+        browser.keys(['Tab']); // Cell 0,0 gets focus
+        expect(browser.$('[class*="column-header-row"] th:nth-child(1)').isFocused()).toBe(true);
+        Terra.validates.element('flowsheet-data-grid-column-highlight-green', { selector: '#terra-flowsheet-data-grid-column-highlight-green' });
+      });
+
+      it('renders a flowsheet data grid with orange and green column highlight', () => {
+        browser.url('/raw/tests/cerner-terra-framework-docs/data-grid/flowsheet-data-grid/multi-column-highlight');
+        browser.keys(['Tab']); // Cell 0,0 gets focus
+        expect(browser.$('[class*="column-header-row"] th:nth-child(1)').isFocused()).toBe(true);
+        Terra.validates.element('flowsheet-data-grid-multi-column-highlight', { selector: '#terra-flowsheet-data-grid-multi-column-highlight' });
+      });
+
+      it('renders a flowsheet data grid sections along with orange column highlight', () => {
+        browser.url('/raw/tests/cerner-terra-framework-docs/data-grid/flowsheet-data-grid/flowsheet-sections-with-col-highlight');
+        Terra.validates.element('flowsheet-data-grid-sections-with-orange-column-highlight', { selector: '#flowsheet-sections-with-col-highlight' });
       });
     });
   });
@@ -155,7 +183,7 @@ Terra.describeViewports('FlowsheetDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(-3, 0);
       releaseShiftKey();
 
-      expect(browser.$('tr.column-header-row th:nth-child(2)').isFocused()).toBe(true);
+      expect(browser.$('[class*="column-header-row"] th:nth-child(2)').isFocused()).toBe(true);
       Terra.validates.element('selection-arrows-range-3-1-to-1-1-focus-0-1', { selector: cellSelectionSelector });
     });