+++ packages/terra-table/src/index.js
@@ -1,12 +1,13 @@
 import Table, { TableConstants, RowSelectionModes } from './Table';
 import GridContext, { GridConstants } from './utils/GridContext';
 import cellShape from './proptypes/cellShape';
-import columnShape from './proptypes/columnShape';
+import columnShape, { ColumnHighlightColor } from './proptypes/columnShape';
 import rowShape from './proptypes/rowShape';
 import sectionShape from './proptypes/sectionShape';
-import validateRowHeaderIndex from './proptypes/validators';
+import { validateRowHeaderIndex, validateAction } from './proptypes/validators';
+import hasColumnActions from './utils/actionsUtils';
 
 export default Table;
 export {
-  GridContext, GridConstants, TableConstants, RowSelectionModes, cellShape, columnShape, rowShape, sectionShape, validateRowHeaderIndex,
+  GridContext, GridConstants, TableConstants, RowSelectionModes, cellShape, columnShape, rowShape, sectionShape, validateRowHeaderIndex, validateAction, hasColumnActions, ColumnHighlightColor,
 };