+++ packages/terra-table/src/subcomponents/Cell.jsx
@@ -16,17 +16,12 @@ import VisuallyHiddenText from 'terra-visually-hidden-text';
 import ColumnContext from '../utils/ColumnContext';
 import GridContext, { GridConstants } from '../utils/GridContext';
 import getFocusableElements from '../utils/focusManagement';
+import { ColumnHighlightColor } from '../proptypes/columnShape';
 import styles from './Cell.module.scss';
 
 const cx = classNames.bind(styles);
 
 const propTypes = {
-  /**
-   * @private
-   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
-   */
-  intl: PropTypes.shape({ formatMessage: PropTypes.func }).isRequired,
-
   /**
    * String identifier of the row in which the Cell will be rendered.
    */
@@ -112,38 +107,72 @@ const propTypes = {
    * With this property the height of the cell will grow to fit the cell content.
    */
   rowMinimumHeight: PropTypes.string,
+
+  /**
+   * A zero-based index indicating which column represents the row header.
+   * Index can be set to -1 if row headers are not required.
+   */
+  rowHeaderIndex: PropTypes.number,
+
+  /**
+   * @private
+   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
+   */
+  intl: PropTypes.shape({ formatMessage: PropTypes.func }).isRequired,
+  /**
+   * @private
+   * Id of the first row in table
+   */
+  firstRowId: PropTypes.string,
+
+  /**
+   * @private
+   * Id of the last row in table
+   */
+  lastRowId: PropTypes.string,
+
+  /**
+   * @private
+   * The color to be used for highlighting a column.
+   */
+  columnHighlightColor: PropTypes.oneOf(Object.values(ColumnHighlightColor)),
 };
 
 const defaultProps = {
+  isMasked: false,
   isRowHeader: false,
   isSelectable: false,
-  isMasked: false,
   sectionId: '',
 };
 
 function Cell(props) {
   const {
-    rowId,
-    columnId,
-    rowIndex,
-    columnIndex,
-    sectionId,
-    tableId,
     ariaLabel,
+    children,
+    columnId,
+    columnIndex,
+    height,
+    intl,
+    isHighlighted,
     isMasked,
-    maskedLabel,
     isRowHeader,
     isSelectable,
     isSelected,
-    isHighlighted,
-    children,
+    maskedLabel,
     onCellSelect,
-    height,
-    intl,
+    rowHeaderIndex,
+    rowId,
+    rowIndex,
     rowMinimumHeight,
+    sectionId,
+    tableId,
+    firstRowId,
+    lastRowId,
+    columnHighlightColor,
   } = props;
 
   const cellRef = useRef();
+
   const theme = useContext(ThemeContext);
   const gridContext = useContext(GridContext);
   const columnContext = useContext(ColumnContext);
@@ -160,30 +189,6 @@ function Cell(props) {
     const focusableElements = getFocusableElements(cellRef.current);
     return focusableElements.length > 0;
   };
-
-  useEffect(() => {
-    if (isGridContext) {
-      setIsInteractable(hasFocusableElements());
-    }
-  }, [isGridContext]);
-
-  /**
-   * Handles the onDeactivate callback for FocusTrap component
-   */
-  const deactivateFocusTrap = () => {
-    setIsFocusTrapEnabled(false);
-    if (gridContext.setCellAriaLiveMessage) {
-      gridContext.setCellAriaLiveMessage(intl.formatMessage({ id: 'Terra.table.resume-navigation' }));
-    }
-  };
-
-  const onMouseDown = ((event) => {
-    if (!isFocusTrapEnabled) {
-      onCellSelect({
-        sectionId, rowId, rowIndex: (rowIndex - 1), columnId, columnIndex, isShiftPressed: event.shiftKey, isMetaPressed: event.metaKey || event.ctrlKey, isCellSelectable: (!isMasked && isSelectable),
-      });
-    }
-  });
 
   /**
    *
@@ -211,6 +216,37 @@ function Cell(props) {
     }
 
     return false;
+  };
+
+  /**
+   * Handles the onDeactivate callback for FocusTrap component
+   */
+  const deactivateFocusTrap = () => {
+    setIsFocusTrapEnabled(false);
+    if (gridContext.setCellAriaLiveMessage) {
+      gridContext.setCellAriaLiveMessage(intl.formatMessage({ id: 'Terra.table.resume-navigation' }));
+    }
+  };
+
+  useEffect(() => {
+    if (isGridContext) {
+      setIsInteractable(hasFocusableElements());
+    }
+  }, [isGridContext]);
+
+  const onMouseDown = (event) => {
+    if (!isFocusTrapEnabled) {
+      onCellSelect({
+        sectionId,
+        rowId,
+        rowIndex: (rowIndex - 1),
+        columnId,
+        columnIndex,
+        isShiftPressed: event.shiftKey,
+        isMetaPressed: event.metaKey || event.ctrlKey,
+        isCellSelectable: (!isMasked && isSelectable),
+      });
+    }
   };
 
   const handleKeyDown = (event) => {
@@ -241,7 +277,14 @@ function Cell(props) {
         case KeyCode.KEY_SPACE:
           if (onCellSelect) {
             onCellSelect({
-              sectionId, rowId, rowIndex: (rowIndex - 1), columnId, columnIndex, isShiftPressed: event.shiftKey, isMetaPressed: event.metaKey || event.ctrlKey, isCellSelectable: (!isMasked && isSelectable),
+              sectionId,
+              rowId,
+              rowIndex: (rowIndex - 1),
+              columnId,
+              columnIndex,
+              isShiftPressed: event.shiftKey,
+              isMetaPressed: event.metaKey || event.ctrlKey,
+              isCellSelectable: (!isMasked && isSelectable),
             });
           }
 
@@ -273,34 +316,24 @@ function Cell(props) {
     cellContent = children;
   }
 
-  const className = cx('cell', {
-    masked: isMasked,
-    pinned: columnIndex < columnContext.pinnedColumnOffsets.length,
-    'last-pinned-column': columnIndex === columnContext.pinnedColumnOffsets.length - 1,
-    selectable: isSelectable && !isMasked,
-    selected: isSelected && !isMasked,
-    highlighted: isHighlighted,
-    blank: !children,
-  }, theme.className);
-
-  const cellLeftEdge = (columnIndex < columnContext.pinnedColumnOffsets.length) ? columnContext.pinnedColumnOffsets[columnIndex] : null;
-
-  const CellTag = isRowHeader ? 'th' : 'td';
-
   // Added to check if rowHeight is defined, it will take precedence. Otherwise the minimum row height would be used.
   const heightProperties = (height) ? {
     height,
   } : { minHeight: rowMinimumHeight };
 
   // eslint-disable-next-line react/forbid-dom-props
-  let cellContentComponent = <div className={cx('cell-content', theme.className)} style={{ ...heightProperties }}>{cellContent}</div>;
+  let cellContentComponent = (<div className={cx('cell-content', theme.className)} style={{ ...heightProperties }}>{cellContent}</div>);
+
   // Render FocusTrap container when within a grid context
   if (isGridContext) {
     cellContentComponent = (
       <FocusTrap
         active={isFocusTrapEnabled}
         focusTrapOptions={{
-          returnFocusOnDeactivate: true, clickOutsideDeactivates: true, escapeDeactivates: false, onDeactivate: deactivateFocusTrap,
+          returnFocusOnDeactivate: true,
+          clickOutsideDeactivates: true,
+          escapeDeactivates: false,
+          onDeactivate: deactivateFocusTrap,
         }}
       >
         {cellContentComponent}
@@ -309,9 +342,31 @@ function Cell(props) {
   }
 
   // Determine table cell header attribute values
+  const cellLeftEdge = (columnIndex < columnContext.pinnedColumnOffsets.length) ? columnContext.pinnedColumnOffsets[columnIndex] : null;
+  const CellTag = isRowHeader ? 'th' : 'td';
+  const columnHeaderId = `${tableId}-${columnId}-headerCell`;
+  const rowHeaderId = !isRowHeader && rowHeaderIndex !== -1 ? `${tableId}-rowheader-${rowId} ` : '';
   const sectionHeaderId = sectionId ? `${tableId}-${sectionId} ` : '';
-  const rowHeaderId = !isRowHeader ? `${tableId}-rowheader-${rowId} ` : '';
-  const columnHeaderId = `${tableId}-${columnId}`;
+
+  let columnHighlight = {};
+  if (columnHighlightColor) {
+    columnHighlight = {
+      [`column-highlight-${columnHighlightColor.toLowerCase()}`]: true,
+      [`first-highlight-${columnHighlightColor.toLowerCase()}`]: rowId === firstRowId,
+      [`last-highlight-${columnHighlightColor.toLowerCase()}`]: rowId === lastRowId,
+    };
+  }
+
+  const className = cx('cell', {
+    masked: isMasked,
+    pinned: columnIndex < columnContext.pinnedColumnOffsets.length,
+    'last-pinned-column': columnIndex === columnContext.pinnedColumnOffsets.length - 1,
+    selectable: isSelectable && !isMasked,
+    selected: isSelected && !isMasked,
+    highlighted: isHighlighted,
+    blank: !children,
+    ...columnHighlight,
+  }, theme.className);
 
   return (
     <CellTag
@@ -322,11 +377,11 @@ function Cell(props) {
       headers={`${sectionHeaderId}${rowHeaderId}${columnHeaderId}`}
       tabIndex={isGridContext ? -1 : undefined}
       className={className}
-      {...(isRowHeader && { scope: 'row', role: 'rowheader' })}
       onMouseDown={onCellSelect ? onMouseDown : undefined}
       onKeyDown={handleKeyDown}
       // eslint-disable-next-line react/forbid-component-props
       style={{ left: cellLeftEdge }}
+      {...(isRowHeader && { scope: 'row', role: 'rowheader' })}
     >
       {cellContentComponent}
       {isInteractable && <VisuallyHiddenText text={intl.formatMessage({ id: 'Terra.table.cell-interactable' })} />}