+++ packages/terra-table/src/subcomponents/Row.jsx
@@ -85,8 +85,21 @@ const propTypes = {
 
   /**
    * A zero-based index indicating which column represents the row header.
+   * Index can be set to -1 if row headers are not required.
    */
   rowHeaderIndex: PropTypes.number,
+
+  /**
+   * @private
+   * Id of the first row in table
+   */
+  firstRowId: PropTypes.string,
+
+  /**
+   * @private
+   * Id of the last row in table
+   */
+  lastRowId: PropTypes.string,
 };
 
 const defaultProps = {
@@ -110,6 +123,8 @@ function Row(props) {
     rowHeaderIndex,
     onCellSelect,
     rowMinimumHeight,
+    firstRowId,
+    lastRowId,
   } = props;
 
   const theme = useContext(ThemeContext);
@@ -171,11 +186,15 @@ function Row(props) {
             isMasked={cellData.isMasked}
             maskedLabel={cellData.maskedLabel}
             isSelectable={cellData.isSelectable}
-            isRowHeader={cellColumnIndex === (rowHeaderIndex + columnIndexOffSet)}
+            isRowHeader={rowHeaderIndex !== -1 && cellColumnIndex === (rowHeaderIndex + columnIndexOffSet)}
             isHighlighted={isHovered || isSelected}
             onCellSelect={onCellSelect}
             height={height}
             rowMinimumHeight={rowMinimumHeight}
+            rowHeaderIndex={rowHeaderIndex}
+            firstRowId={firstRowId}
+            lastRowId={lastRowId}
+            columnHighlightColor={displayedColumns[cellColumnIndex].columnHighlightColor}
           >
             {cellData.content}
           </Cell>