+++ packages/terra-table/src/subcomponents/Section.jsx
@@ -89,6 +89,7 @@ const propTypes = {
 
   /**
    * A zero-based index indicating which column represents the row header.
+   * Index can be set to -1 if row headers are not required.
    */
   rowHeaderIndex: PropTypes.number,
 
@@ -101,6 +102,21 @@ const propTypes = {
    * Function that is called when a collapsible section is selected. Parameters: `onSectionSelect(sectionId)`
    */
   onSectionSelect: PropTypes.func,
+  /**
+   * Bounding container for the table, will use window if no value provided.
+   */
+  boundingRef: PropTypes.func,
+  /**
+   * @private
+   * Id of the first row in table
+  */
+  firstRowId: PropTypes.string,
+
+  /**
+    * @private
+    * Id of the last row in table
+    */
+  lastRowId: PropTypes.string,
 };
 
 const defaultProps = {
@@ -128,6 +144,9 @@ function Section(props) {
     rows,
     onSectionSelect,
     rowMinimumHeight,
+    boundingRef,
+    firstRowId,
+    lastRowId,
   } = props;
 
   const theme = useContext(ThemeContext);
@@ -136,6 +155,8 @@ function Section(props) {
   const isGridContext = gridContext.role === GridConstants.GRID;
 
   const hasSectionButton = isCollapsible && onSectionSelect;
+  const boundedWidth = isCollapsible && boundingRef && boundingRef.current ? boundingRef.current.clientWidth - 50 : null;
+  const titlePosition = boundingRef ? { isTitleSticky: true } : { isTitleFixed: true };
 
   const handleClick = useCallback(() => {
     onSectionSelect(id);
@@ -164,8 +185,9 @@ function Section(props) {
               className={cx('section-header')}
               text={text}
               isOpen={hasSectionButton ? !isCollapsed : undefined}
-              isTitleFixed
               onClick={hasSectionButton ? handleClick : undefined}
+              boundedWidth={boundedWidth}
+              {...titlePosition}
             />
           </th>
         </tr>
@@ -194,6 +216,8 @@ function Section(props) {
             isSelected={row.isSelected}
             isTableStriped={isTableStriped}
             rowMinimumHeight={rowMinimumHeight}
+            firstRowId={firstRowId}
+            lastRowId={lastRowId}
           />
         ))}
       </tbody>