+++ packages/terra-tabs/src/common-tabs/_Tabs.jsx
@@ -160,15 +160,19 @@ class Tabs extends React.Component {
     const currTab = this.props.tabData.find((tab) => tab.isSelected === true);
     const prevtabKeys = [];
     const prevtabLabels = [];
+    const prevtabIcons = [];
     prevProps.tabData.forEach(child => {
       prevtabKeys.push(child.id);
       prevtabLabels.push(child.label);
+      prevtabIcons.push(child.icon);
     });
     const curtabKeys = [];
     const curtabLabels = [];
+    const curtabIcons = [];
     this.props.tabData.forEach(child => {
       curtabKeys.push(child.id);
       curtabLabels.push(child.label);
+      curtabIcons.push(child.icon);
     });
 
     // Allow dynamic addition of tabs.
@@ -180,9 +184,11 @@ class Tabs extends React.Component {
       for (let i = 0; i < curtabKeys.length; i += 1) {
         const prevKey = prevtabKeys[i];
         const prevLabel = prevtabLabels[i];
+        const prevIcon = prevtabIcons[i];
         const curKey = curtabKeys[i];
         const curLabel = curtabLabels[i];
-        if (prevKey !== curKey || prevLabel !== curLabel) {
+        const curIcon = curtabIcons[i];
+        if (prevKey !== curKey || prevLabel !== curLabel || prevIcon !== curIcon) {
           isTabEqual = true;
         }
       }