+++ packages/terra-date-input/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.54.0 - (March 8, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.53.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.52.0 - (February 22, 2024)
 
 * Changed