+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+## 4.111.0 - (March 8, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 4.110.0 - (March 5, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 4.109.0 - (March 1, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 4.108.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 4.107.0 - (February 22, 2024)
 
 * Added