+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+## 3.107.0 - (March 8, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 3.106.0 - (March 5, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 3.105.0 - (March 1, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 3.104.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 3.103.0 - (February 22, 2024)
 
 * Changed