+++ packages/terra-pills/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 1.21.0 - (March 8, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.20.0 - (March 5, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.19.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.18.0 - (January 22, 2024)
 
 * Changed