+++ packages/terra-tabs/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+## 7.20.1 - (March 8, 2024)
+
+* Fixed
+  * Fixes component re-render with updated icon.
+
+## 7.20.0 - (March 5, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 7.19.0 - (March 1, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 7.18.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 7.17.0 - (February 7, 2024)
 
 * Changed
@@ -50,7 +70,7 @@
 
 * Added
   * Added Vertical orientation for `terra-tabs`.
-  
+
 * Changed
   * Removed additional screen reader phrase and fixed delete tab hint.
 
@@ -67,7 +87,7 @@
   * Fixes large tab content display with scrollable view.
 
 * Fixed
-  * Fixed Axe core violations for close button added on 
+  * Fixed Axe core violations for close button added on
     terra-tabs on both face-up tab and hidden tab.
   * Fixed issue in screen reader instruction for hidden tabs   where add tab instruction is announced even in absence of add button.
 
@@ -136,7 +156,7 @@
   * Introduced `variant` prop to switch between framework and workspace styles
 
 * Changed
-  * Updated framework tabs to consume common tabs 
+  * Updated framework tabs to consume common tabs
   * Updated styles to accommodate both framework and workspace styles
 
 ## 6.72.0 - (July 19, 2023)