+++ packages/terra-time-input/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 4.65.0 - (March 8, 2024)
+
+* Fixed
+  * Time input displays incorrectly in firefox.
+
+## 4.64.0 - (February 28, 2024)
+
 * Fixed
   * Fixed error message reading issue.